logsec rework
