+++ ledfx/effects/game_of_life.py
@@ -164,8 +164,8 @@ class GameOfLifeVisualiser(Twod):
         if self.test:
             self.draw_test(self.m_draw)
 
-        self.check_board_health(self.current_time)
-        self.step_board_if_time_elapsed(self.current_time)
+        self.check_board_health(self.logsec.current_time)
+        self.step_board_if_time_elapsed(self.logsec.current_time)
         self.update_image_with_board()
 
     def check_board_health(self, current_time):
